# docs/zooearly-ai-openapi.yaml
openapi: 3.0.3
info:
  title: ZooEarly AI API
  version: 2.0.0
  description: |
    쥬얼리 (ZooEarly) — AI API 명세서 (v1.0 · 2026-08-21)

    React Native 앱 ↔ API Gateway ↔ FastAPI Inference Server(STT / LLM / TTS → OpenAI API)

    - `/ai/*` 는 릴레이 전용: 요청 검증 + 전달 + 에러 통일만 한다. 무상태.
    - `/albums` 만 다르다 — 게이트웨이가 MySQL 에 직접 남기는 동화 앨범이다(2026-08-27).
    - 상태(대화 이력, 목표 문장, 시나리오)는 전부 앱이 갖고 있으며 매 요청에 실어 보낸다.
    - FastAPI 내부 경로는 접두사만 다르다: `/api/v1/ai/chat` → `/ai/chat`
    - 어떤 에러도 아이 화면에 "오류"로 표시하지 않는다. 전부 "괜찮아, 다시 해볼까?"로 폴백한다.
servers:
  # ★ 아래 주소에 /api/v1 이 포함돼 있다. paths 는 그 뒤에 붙는 부분만 적은 것이다.
  #   예) /ai/story  ->  https://<호스트>/api/v1/ai/story
  #
  #   앱의 base URL 을 /api/v1 없이 호스트까지만 넣으면 전부 404다.
  #   실제로 그 사고가 있었다(2026-08-25) — paths 의 /ai/... 만 보고
  #   base URL 을 호스트로만 설정해서 모든 호출이 404·CORS 에러로 보였다.
  - url: https://zooearly-gw.jollyhill-992be81c.koreacentral.azurecontainerapps.io/api/v1
    description: 배포된 게이트웨이 (Azure Container Apps, koreacentral)
  - url: http://localhost:8080/api/v1
    description: 로컬 개발 (gradlew bootRun)
...
paths:
...
  /albums:
    post:
      tags: [Album]
      operationId: saveAlbum
      summary: 동화를 앨범에 남긴다
      description: |
        하루를 마치며 만든 동화를 남긴다. 다음 날에도 다시 읽을 수 있게 된다.

        **앱은 동화를 화면에 띄운 뒤에 부른다.** 저장을 기다렸다 보여주면 저장이
        늦거나 실패하는 날 아이가 동화를 아예 못 본다. 실패해도 화면을 막지 않는다.

        **삽화는 보내지 않는다.** 앱 번들의 정적 그림이고 장면의 `category` 하나로
        결정된다 — 이미지를 저장하면 동화마다 같은 파일을 복제하면서, 나중에 그림을
        고쳐도 옛 앨범만 낡은 그림으로 남는다.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AlbumSaveRequest'
      responses:
        '200':
          description: 저장됨
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessEnvelope'
                  - type: object
                    properties:
                      data:
                        type: object
                        required: [id]
                        properties:
                          id: { type: integer, format: int64, example: 12 }
        '400': { $ref: '#/components/responses/BadRequest' }

    get:
      tags: [Album]
      operationId: listAlbums
      summary: 내 동화 목록 (최신순)
      description: |
        표지를 그리는 데 필요한 것만 준다. `categories` 는 목록에서 **보석 줄**을
        그리는 데 쓴다 — 글자를 아직 못 읽는 아이도 색이 다른 보석 네 개를 보고
        그날을 알아본다.
      parameters:
        - name: childId
          in: query
          required: true
          schema: { $ref: '#/components/schemas/ChildId' }
      responses:
        '200':
          description: 목록
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessEnvelope'
                  - type: object
                    properties:
                      data:
                        type: array
                        items: { $ref: '#/components/schemas/AlbumSummary' }
        '400': { $ref: '#/components/responses/BadRequest' }
  /albums/{id}:
    get:
      tags: [Album]
      operationId: getAlbum
      summary: 동화 한 편
      description: |
        앱이 이걸 받아 동화 화면을 그대로 다시 그린다.

        **`childId` 를 반드시 함께 보낸다.** `id` 만으로 찾지 않는 이유는 인증이
        없기 때문이다 — 번호를 하나씩 올려보는 것만으로 남의 동화를 읽을 수 있다.

        주인이 아닐 때 `403` 이 아니라 `404` 를 준다. `403` 은 "있긴 있다"를
        알려주는 셈이라 번호를 훑어 남의 동화가 몇 편인지 셀 수 있다.
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer, format: int64 }
        - name: childId
          in: query
          required: true
          schema: { $ref: '#/components/schemas/ChildId' }
      responses:
        '200':
          description: 동화 한 편
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessEnvelope'
                  - type: object
                    properties:
                      data: { $ref: '#/components/schemas/AlbumDetail' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '404':
          description: 없거나 내 것이 아니다
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorEnvelope' }
components:
  schemas:
...
    ChildId:
      type: string
      format: uuid
      pattern: '^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$'
      example: a355d4f5-f7ac-4b74-8733-b28236b270ac
      description: |
        기기가 처음 켜질 때 만든 UUIDv4. 로그인이 없으므로 **이것이 "누구인가"의 전부다.**

        닉네임을 키로 쓰지 않는 이유는 셋이다 — 겹치고(같은 반에 "지우"가 둘),
        바뀌고(메뉴에서 이름 수정), 아이 이름이 모든 조회 주소에 실린다.

        앱은 `localStorage['zooearly.childId']` 에 프로필과 **따로** 보관한다.
        날짜가 바뀌어 진행도가 비워질 때 앨범까지 잃으면 안 되기 때문이다.
        다만 "처음부터 플레이하기"(기기를 다음 아이에게 넘기는 동선)에서는 새로 만든다.
    AlbumScene:
      type: object
      required: [category, subtitle, narration]
      properties:
        category:
          type: string
          enum: [school_arrival, class, lunch, school_departure]
          description: |
            **삽화를 고르는 유일한 키.** 앱이 이 값으로 번들의 정적 그림을 찾는다.
            그래서 이미지를 저장하지 않는다.
        subtitle: { type: string, maxLength: 120, example: 학교 오는 길 }
        opening: { type: string, maxLength: 2000, example: 아침에 }
        quote:
          type: string
          maxLength: 500
          nullable: true
          description: 아이가 실제로 고른 말. 화면에 인용 블록으로 뜬다. 고르지 않고 넘어간 장면은 null.
          example: 안녕! 우리 같이 놀자!
        narration: { type: string, maxLength: 4000, example: 지우가 학교에 왔어요. }
    AlbumSaveRequest:
      type: object
      required: [childId, nickname, title, scenes]
      properties:
        childId: { $ref: '#/components/schemas/ChildId' }
        nickname:
          type: string
          maxLength: 40
          example: 지우
          description: 만들 때의 이름. **표시용이지 키가 아니다** — 나중에 이름을 바꿔도 옛 표지는 그대로다.
        title: { type: string, maxLength: 120, example: 지우가 들려준 오늘 }
        scenes:
          type: array
          minItems: 1
          items: { $ref: '#/components/schemas/AlbumScene' }
    AlbumSummary:
      type: object
      required: [id, title, nickname, createdAt, categories]
      properties:
        id: { type: integer, format: int64, example: 12 }
        title: { type: string, example: 지우가 들려준 오늘 }
        nickname: { type: string, example: 지우 }
        createdAt: { type: string, format: date-time, example: '2026-08-27T01:10:51.313624Z' }
        categories:
          type: array
          description: 목록에서 보석 줄을 그리는 데 쓴다. 장면 순서대로.
          items:
            type: string
            enum: [school_arrival, class, lunch, school_departure]
    AlbumDetail:
      type: object
      required: [id, title, nickname, createdAt, scenes]
      properties:
        id: { type: integer, format: int64 }
        title: { type: string }
        nickname:
          type: string
          description: 그때 그 이름. 지금 프로필과 다를 수 있다.
        createdAt: { type: string, format: date-time }
        scenes:
          type: array
          items: { $ref: '#/components/schemas/AlbumScene' }
    SuccessEnvelope:
      type: object
      required: [success, data]
      properties:
        success:
          type: boolean
          enum: [true]
        data:
          type: object
    ErrorEnvelope:
      type: object
      required: [success, error]
      properties:
        success:
          type: boolean
          enum: [false]
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              enum:
                - INVALID_PARAMETER
                - UNSUPPORTED_AUDIO_FORMAT
                - AUDIO_TOO_LARGE
                - PAYLOAD_TOO_LARGE
                - STT_FAILED
                - RATE_LIMITED
                - AI_SERVER_ERROR
                - AI_TIMEOUT
                - INTERNAL_ERROR
            message:
              type: string
              description: 개발자용. 아이 화면에 띄우지 않는다
            field:
              type: string
              nullable: true
              description: 문제가 된 파라미터명. 검증 에러에만 채워진다
  responses:
    BadRequest:
      description: "검증 실패 (Gateway) — INVALID_PARAMETER / UNSUPPORTED_AUDIO_FORMAT / AUDIO_TOO_LARGE"
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
          example:
            success: false
            error: { code: INVALID_PARAMETER, message: "audio 필드가 없습니다.", field: audio }
...